Suppressed warning in maven2 environment “Using platform encoding (UTF-8 actually) to copy filtered resources, i.e. build is platform dependent!” using the properties.
Maven2 warning is because of the fact that maven2 plugins uses platform default encoding if no explicit file encoding has been provided in the plugin configuration. The usages of the platform encoding makes the build platform dependent hence the maven2 is bound to raise an alarm as build artifact/resources become irreproducible.